+++ include/omw/windows/string.h
@@ -1,6 +1,6 @@
 /*
 author          Oliver Blaser
-date            17.12.2023
+date            30.12.2023
 copyright       MIT - Copyright (c) 2023 Oliver Blaser
 */
 
@@ -25,6 +25,13 @@ namespace omw
 
         //! \name String Conversion
         /// @{
+        std::wstring u8tows(const char* src);
+        static inline std::wstring u8tows(const std::string& src) { return u8tows(src.c_str()); }
+
+        std::string wstou8(const wchar_t* src);
+        static inline std::string wstou8(const std::wstring& src) { return wstou8(src.c_str()); }
+
+        // TODO deprecated, remove
         size_t utf8_to_wstr(const char* src, wchar_t* dest, size_t destSize);
         size_t utf8_to_wstr(const char* src, wchar_t* dest, size_t destSize, omw::windows::ErrorCode& ec);
         static inline size_t utf8_to_wstr(const std::string& src, wchar_t* dest, size_t destSize) { return utf8_to_wstr(src.c_str(), dest, destSize); }